split functionallity into multiple classes
