--- src/client/datascience/export/exportBase.ts
@@ -1,13 +1,0 @@
-import { injectable } from 'inversify';
-import { Uri } from 'vscode';
-
-export const IExport = Symbol('IExport');
-export interface IExport {
-    export(source: Uri, target: Uri): Promise<void>;
-}
-
-@injectable()
-export abstract class ExportBase implements IExport {
-    // tslint:disable-next-line: no-empty
-    public async export(_source: Uri, _target: Uri): Promise<void> {}
-}
