+++ src/client/datascience/export/exportToPython.ts
@@ -3,18 +3,16 @@ import { Uri, ViewColumn } from 'vscode';
 import { IDocumentManager } from '../../common/application/types';
 import { IFileSystem } from '../../common/platform/types';
 import { INotebookEditorProvider, INotebookImporter } from '../types';
-import { ExportBase } from './exportBase';
+import { IExport } from './exportManager';
 
 @injectable()
-export class ExportToPython extends ExportBase {
+export class ExportToPython implements IExport {
     constructor(
         @inject(INotebookEditorProvider) protected readonly notebookEditorProvider: INotebookEditorProvider,
         @inject(IFileSystem) protected readonly fileSystem: IFileSystem,
         @inject(IDocumentManager) protected readonly documentManager: IDocumentManager,
         @inject(INotebookImporter) protected readonly importer: INotebookImporter
-    ) {
-        super();
-    }
+    ) {}
 
     public async export(source: Uri): Promise<void> {
         const contents = await this.importer.importFromFile(
